+++ tests/integration/test_board_use.gd
@@ -79,24 +79,25 @@ func test_card_hand_drop_recovery():
 			"Card dragged back in hand remains in hand")
 
 func test_card_drag_block_by_board_borders():
-	cards[4]._on_Card_mouse_entered()
+	var card = cards[4]
+	card._on_Card_mouse_entered()
 	click_card(cards[4])
 	yield(yield_for(0.5), YIELD) # Wait to allow dragging to start
-	board._UT_interpolate_mouse_move(Vector2(-100,100),cards[0].global_position)
+	board._UT_interpolate_mouse_move(Vector2(-100,100),card.global_position)
 	yield(yield_for(0.4), YIELD)
-	assert_almost_eq(Vector2(0, 100),cards[4].global_position,Vector2(2,2),
+	assert_almost_eq(Vector2(0, 100),card.global_position,Vector2(2,2),
 			"Dragged outside left viewport borders stays inside viewport")
 	board._UT_interpolate_mouse_move(Vector2(1300,300))
 	yield(yield_for(0.4), YIELD)
-	assert_almost_eq(Vector2(1220, 300),cards[4].global_position,Vector2(2,2),
+	assert_almost_eq(Vector2(1220, 300),card.global_position,Vector2(2,2),
 			"Dragged outside right viewport borders stays inside viewport")
 	board._UT_interpolate_mouse_move(Vector2(800,-100))
 	yield(yield_for(0.4), YIELD)
-	assert_almost_eq(Vector2(800, 0),cards[4].global_position,Vector2(2,2),
+	assert_almost_eq(Vector2(800, 0),card.global_position,Vector2(2,2),
 			"Dragged outside top viewport borders stays inside viewport")
 	board._UT_interpolate_mouse_move(Vector2(500,800))
 	yield(yield_for(0.4), YIELD)
-	assert_almost_eq(Vector2(500, 624),cards[4].global_position,Vector2(2,2),
+	assert_almost_eq(Vector2(500, 624),card.global_position,Vector2(2,2),
 			"Dragged outside bottom viewport borders stays inside viewport")
 
 
@@ -115,3 +116,12 @@ func test_fast_card_table_drop():
 	yield(yield_for(0.6), YIELD)
 	assert_almost_eq(Vector2(1000, 300),cards[0].global_position,Vector2(2,2),
 			"Card not dragged with mouse after dropping on table")
+
+func test_board_to_board_move():
+	var card: Card
+	card = cards[0]
+	yield(table_move(card, Vector2(100,200)), "completed")
+	card.card_rotation = 90
+	yield(drag_drop(card, Vector2(800,200)), 'completed')
+	assert_eq(90.0,card.get_node("Control").rect_rotation,
+			"Card should stay in the same rotation when moved around the board")