+++ tests/integration/test_manipulation_buttons.gd
@@ -8,6 +8,7 @@ func before_each():
 	yield(yield_for(1), YIELD)
 
 func test_manipulation_buttons_not_messing_hand_focus():
+	cfc.hand_use_oval_shape = false
 	cards[0]._on_Card_mouse_entered()
 	yield(yield_for(0.05), YIELD)
 	cards[0]._on_Card_mouse_exited()
@@ -16,10 +17,7 @@ func test_manipulation_buttons_not_messing_hand_focus():
 	cards[0]._on_button_mouse_exited()
 	cards[0]._on_Card_mouse_entered()
 	yield(yield_to(cards[0].get_node('Tween'), "tween_all_completed", 1), YIELD)
-	if cfc.hand_use_oval_shape:
-		pass
-	else:
-		assert_almost_eq(Vector2(47.5, -240),cards[0].position,Vector2(2,2),
-				"Card dragged in correct global position")
-		assert_almost_eq(Vector2(1.5, 1.5),cards[0].scale,Vector2(0.1,0.1),
-				"Card has correct scale")
+	assert_almost_eq(Vector2(47.5, -240),cards[0].position,Vector2(2,2),
+			"Card dragged in correct global position")
+	assert_almost_eq(Vector2(1.5, 1.5),cards[0].scale,Vector2(0.1,0.1),
+			"Card has correct scale")