+++ changelog.md
@@ -1,11 +1,13 @@
 # Changelog
 
-## 0.12 (Ongoing)
+## 0.12
 
 * Scipting Engine target finding now supports tutoring cards from deck
 * Scipting Engine target finding now supports affecting multiple cards from the board by filtering their properties
 * Scripting Engine now supports cards with multiple choices for abilities.
-* Hand shape can now be oval (default). Contribution by [@vmjcv](https://github.com/vmjcv)
+* Scripting Engine now supports cards with optional abilities
+* Hand shape can now be oval (default) - Contribution by [@vmjcv](https://github.com/vmjcv).
+* Font-size will decrease when it amount of chars is too large, in order to stay within the defined label rectangle - Contribution by [@vmjcv](https://github.com/vmjcv).
 
 ## 0.11